+++ includes/features/class-wpcli.php
@@ -214,7 +214,7 @@ class Wpcli {
 	 * <enable|disable>
 	 * : The action to take.
 	 *
-	 * <gc|analytics>
+	 * <gc|analytics|metrics>
 	 * : The setting to change.
 	 *
 	 * [--yes]
@@ -247,6 +247,10 @@ class Wpcli {
 						Option::network_set( 'gc', true );
 						$this->success( 'garbage collection is now activated.', '', $stdout );
 						break;
+					case 'metrics':
+						Option::network_set( 'metrics', true );
+						$this->success( 'metrics collation is now activated.', '', $stdout );
+						break;
 					default:
 						$this->error( 1, $stdout );
 				}
@@ -262,6 +266,11 @@ class Wpcli {
 						\WP_CLI::confirm( 'Are you sure you want to deactivate garbage collection?', $assoc_args );
 						Option::network_set( 'gc', false );
 						$this->success( 'garbage collection is now deactivated.', '', $stdout );
+						break;
+					case 'metrics':
+						\WP_CLI::confirm( 'Are you sure you want to deactivate metrics collation?', $assoc_args );
+						Option::network_set( 'metrics', false );
+						$this->success( 'metrics collation is now deactivated.', '', $stdout );
 						break;
 					default:
 						$this->error( 1, $stdout );